ci(update): upgrade GitHub Actions and Go version
This commit updates the GitHub Actions used in the CI workflow to their newer versions and upgrades the Go version specified for the setup. Specifically, `actions/checkout` has been upgraded from v2 to v4, and `actions/setup-go` from v2 to v5. Additionally, the Go version has been updated from 1.18 to 1.22.

Upgrading these actions ensures compatibility with the latest features and improvements offered by GitHub Actions, and using the newer Go version allows taking advantage of the latest language improvements and performance enhancements. This update is crucial for maintaining the efficiency and reliability of the CI process.